Revert "Seanaye/feat/push notif for signal email (#3040)"
This reverts commit 82618095b9de0a25548e075ea7a6c7c77ec339ac.
