+++ rust/cloud-storage/email/src/outbound/email_pg_repo/dynamic/filters.rs
@@ -1,6 +1,5 @@
 use super::SqlFragment;
 use crate::domain::models::{PreviewView, PreviewViewStandardLabel};
-use email_importance::build_importance_condition;
 use filter_ast::Expr;
 use item_filters::ast::email::{Email, EmailLiteral};
 use recursion::CollapsibleExt;
@@ -57,6 +56,44 @@ fn build_email_match(preamble: &str, email: &Email) -> SqlFragment {
     }
 }
 
+fn build_sender_importance_override_filter(is_important: bool) -> SqlFragment {
+    let importance_literal = if is_important { "TRUE" } else { "FALSE" };
+    let opposite_importance_literal = if is_important { "FALSE" } else { "TRUE" };
+
+    SqlFragment::raw(format!(
+        r#"(
+                    EXISTS (
+                        SELECT 1
+                        FROM email_contacts sender_c
+                        JOIN email_filters ef
+                          ON ef.link_id = m.link_id
+                         AND ef.email_address IS NOT NULL
+                         AND LOWER(ef.email_address) = LOWER(sender_c.email_address)
+                        WHERE sender_c.id = m.from_contact_id
+                          AND ef.is_important = {importance_literal}
+                    )
+                    OR EXISTS (
+                        SELECT 1
+                        FROM email_contacts sender_c
+                        JOIN email_filters ef
+                          ON ef.link_id = m.link_id
+                         AND ef.email_domain IS NOT NULL
+                         AND LOWER(ef.email_domain) = LOWER(split_part(sender_c.email_address, '@', 2))
+                        WHERE sender_c.id = m.from_contact_id
+                          AND ef.is_important = {importance_literal}
+                          AND NOT EXISTS (
+                              SELECT 1
+                              FROM email_filters ef_addr
+                              WHERE ef_addr.link_id = m.link_id
+                                AND ef_addr.email_address IS NOT NULL
+                                AND LOWER(ef_addr.email_address) = LOWER(sender_c.email_address)
+                                AND ef_addr.is_important = {opposite_importance_literal}
+                          )
+                    )
+                )"#,
+    ))
+}
+
 pub(super) fn build_message_email_filter(ast: &Expr<EmailLiteral>) -> SqlFragment {
     let fragment = ast.collapse_frames(|frame| match frame {
         filter_ast::ExprFrame::And(a, b) => SqlFragment::and(a, b),
@@ -102,17 +139,77 @@ pub(super) fn build_message_email_filter(ast: &Expr<EmailLiteral>) -> SqlFragmen
         ),
 
         filter_ast::ExprFrame::Literal(EmailLiteral::Importance(true)) => {
-            let mut builder = sqlx::QueryBuilder::new(
-                "(\n                NOT EXISTS (\n                    SELECT 1 FROM email_message_labels ml\n                    JOIN email_labels l ON ml.label_id = l.id\n                    WHERE ml.message_id = m.id\n                    AND l.name = 'TRASH'\n                )\n                AND ",
+            let mut f = SqlFragment::raw(
+                r#"(
+                NOT EXISTS (
+                    SELECT 1 FROM email_message_labels ml
+                    JOIN email_labels l ON ml.label_id = l.id
+                    WHERE ml.message_id = m.id
+                    AND l.name = 'TRASH'
+                )
+                AND (
+                    "#,
             );
-            build_importance_condition(true, &mut builder);
-            builder.push(")");
-            SqlFragment::raw(builder.sql())
+            f.extend(build_sender_importance_override_filter(true));
+            f.push_raw(
+                r#"
+                    OR (
+                        NOT "#,
+            );
+            f.extend(build_sender_importance_override_filter(false));
+            f.push_raw(
+                r#"
+                        AND (
+                            m.is_draft = TRUE
+                            OR EXISTS (
+                                SELECT 1 FROM email_message_labels ml
+                                JOIN email_labels l ON ml.label_id = l.id
+                                WHERE ml.message_id = m.id
+                                AND l.name IN ('CATEGORY_PERSONAL', 'SENT', 'DRAFT')
+                            )
+                            OR NOT EXISTS (
+                                SELECT 1 FROM email_message_labels ml
+                                JOIN email_labels l ON ml.label_id = l.id
+                                WHERE ml.message_id = m.id
+                                AND l.name IN ('CATEGORY_UPDATES', 'CATEGORY_PROMOTIONS', 'CATEGORY_SOCIAL', 'CATEGORY_FORUMS')
+                            )
+                        )
+                    )
+                )
+            )"#,
+            );
+            f
         }
         filter_ast::ExprFrame::Literal(EmailLiteral::Importance(false)) => {
-            let mut builder = sqlx::QueryBuilder::new("");
-            build_importance_condition(false, &mut builder);
-            SqlFragment::raw(builder.sql())
+            let mut f = SqlFragment::raw(
+                r#"(
+                "#,
+            );
+            f.extend(build_sender_importance_override_filter(false));
+            f.push_raw(
+                r#"
+                OR (
+                    NOT "#,
+            );
+            f.extend(build_sender_importance_override_filter(true));
+            f.push_raw(
+                r#"
+                    AND NOT EXISTS (
+                        SELECT 1 FROM email_message_labels ml
+                        JOIN email_labels l ON ml.label_id = l.id
+                        WHERE ml.message_id = m.id
+                        AND l.name IN ('CATEGORY_PERSONAL', 'SENT', 'DRAFT')
+                    )
+                    AND EXISTS (
+                        SELECT 1 FROM email_message_labels ml
+                        JOIN email_labels l ON ml.label_id = l.id
+                        WHERE ml.message_id = m.id
+                        AND l.name IN ('CATEGORY_UPDATES', 'CATEGORY_PROMOTIONS', 'CATEGORY_SOCIAL', 'CATEGORY_FORUMS')
+                    )
+                )
+            )"#,
+            );
+            f
         }
         filter_ast::ExprFrame::Literal(EmailLiteral::NotificationDone(_)) => {
             SqlFragment::raw("TRUE")
