--- rust/cloud-storage/email_importance/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-edition = "2024"
-name = "email_importance"
-publish = false
-version = "0.1.0"
-
-[dependencies]
-anyhow = { workspace = true }
-sqlx = { workspace = true }
-tokio = { workspace = true }
-tracing = { workspace = true }
-uuid = { workspace = true }
-
-[dev-dependencies]
-macro_db_migrator = { path = "../macro_db_migrator" }
