+++ rust/cloud-storage/justfile
@@ -96,7 +96,6 @@ setup_test_envs:
     echo "DATABASE_URL={{ DATABASE_URL }}" >> ./entity_access_management/.env
     echo "DATABASE_URL={{ DATABASE_URL }}" >> ./mobile_welcome_email_db_client/.env
     echo "DATABASE_URL={{ DATABASE_URL }}" >> ./entity_access_db_utils/.env
-    echo "DATABASE_URL={{ DATABASE_URL }}" >> ./email_importance/.env
 
 setup_macrodb:
     just macro_db_client/create_db
